+++ engine/client/gl_vidnt.h
@@ -1,0 +1,21 @@
+
+
+#define VID_NOMODE -2.0f
+#define VID_AUTOMODE	"-1"
+#define VID_DEFAULTMODE	2.0f
+#define DISP_CHANGE_BADDUALVIEW	-6 // MSVC 6.0 doesn't
+#define WINDOW_NAME			"Xash Window" // Half-Life
+
+extern int num_vidmodes;
+
+
+
+typedef struct vidmode_s
+{
+	const char	*desc;
+	int		width;
+	int		height;
+	qboolean		wideScreen;
+} vidmode_t;
+
+extern vidmode_t vidmode[];
